fix: refactor LOC checks to use Python standard library and enhance flow close target tests
