+++ scripts/hooks/check-test-file-loc.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 # Check per-file LOC ceiling for test files
-# Reads limits from config/settings.yaml
+# Reads limits from config/settings.yaml without uv
 #
 # Limits:
 #   default: 200 lines (most test files should fit)
@@ -16,8 +16,43 @@
 set -e
 
 get_limit() {
-  PYTHONPATH=src uv run python -m vibe3.config.get \
-    "$1" -c config/settings.yaml --quiet 2>/dev/null || echo "$2"
+  python3 - "$1" "${2:-}" <<'PY'
+import re
+import sys
+from pathlib import Path
+
+
+def parse_config(path: str) -> dict[str, str]:
+  values: dict[str, str] = {}
+  stack: list[tuple[int, str]] = []
+
+  for raw_line in Path(path).read_text().splitlines():
+    stripped = raw_line.strip()
+    if not stripped or stripped.startswith("#"):
+      continue
+
+    indent = len(raw_line) - len(raw_line.lstrip(" "))
+    while stack and stack[-1][0] >= indent:
+      stack.pop()
+
+    match = re.match(r"([A-Za-z0-9_]+):(?:\s*(.*))?$", stripped)
+    if not match:
+      continue
+
+    key, value = match.groups()
+    if not value:
+      stack.append((indent, key))
+      continue
+
+    path_key = ".".join([item[1] for item in stack] + [key])
+    values[path_key] = value.split("#", 1)[0].strip().strip('"').strip("'")
+
+  return values
+
+
+config = parse_config("config/settings.yaml")
+print(config.get(sys.argv[1], sys.argv[2]))
+PY
 }
 
 # Get limits (unified for all test files)
